Lower negative cache TTL for coredns
This change updates the coredns config map to lower the TTL for caching
negative responses from 30 seconds down to 5 seconds. This will improve
the response time for cases where a hostname lookup might occur before a
given pod is created, resulting in a negative entry being cached and
preventing pods from resolving the name for 30 seconds afterwards.

The default cache size of 9984 items is unchanged, but must be
explicitly defined in this configuration.

The change also adds coredns to the upgrade-k8s-networking playbook to
ensure that changes to coredns can be automatically deployed by
sysinv-conductor.

Closes-Bug: 1906870

Change-Id: I7b44358508c32c8ca4e58b1e69d6232f1a61bfcf
Signed-off-by: Cole Walker <cole.walker@windriver.com>