B&R: Fix restore of ceph using mds
Calling /etc/init.d/ceph without specifying a component applies an
action to all components.
After cephfs_platform_integ topic merged, ceph mds component is also
used.
Mds requires a monitor to be up before starting.
Due to the fact that the ceph script was called without specifying the
components mds the start action was applied to mds also and it was
applied before starting the monitor, resulting in failure.

The fix is to specify the monitor and osd components.
Mds will be start when ceph is restarted a few tasks later.

Plus minor ansible changes:
- Use failed_when instead of ignore_errors
- Renamed task

Closes-Bug: 1912488
Signed-off-by: Dan Voiculeasa <dan.voiculeasa@windriver.com>
Change-Id: I5388ddd32923bbd9acb205fa2d2adbc083346631