B&R: Fix backup hangout on IPv6 systems
IPv6 addreses must be enclosed in square brackets when they are
followed by a port number.

At the "[backup/backup-system : Create etcd snapshot]"
step, the etcd IPv6 endpoint is not wrapped in square brackets,
so the command hangs indefinitely.

We fix this by using the 'ipwrap' ansible filter which will
wrap the address in [] brackets if it's an IPv6 one.

Closes-bug: 1916053
Change-Id: If40ed59f4e44c9f877aaefe87f6211a3e83ddfee
Signed-off-by: Mihnea Saracin <Mihnea.Saracin@windriver.com>