Fix bootstrap playbook when initializing kubernetes
The etcd endpoint in the kubedeadm file is different than
the endpoint defined in the etcd config file.
This is because in the kubeadm file, the etcd endpoint is equal to
the 'cluster_floating_address' variable. And in the etcd config
file the 'default_cluster_host_start_address' variable is used.
These 2 variables can be different when the
'cluster_host_start_address' variable defined in the localhost.yml
differs from the first address of the 'cluster_host_subnet'.

The solution is to use 'cluster_floating_address' in both cases
because this variable is defined in the following way:

cluster_floating_address: "{{ address_pairs['cluster_host']['start'] }}"
will be different than the 'default_cluster_host_start_address': "{{
(cluster_host_subnet | ipaddr(1)).split('/')[0] }

So it will use 'default_cluster_host_start_address' when the
'cluster_host_start_address' is not defined.

Closes-Bug: 1918130
Change-Id: I8fecc1e5e54b5a9a9a72a54c069f79f5f2d434ba
Signed-off-by: Mihnea Saracin <Mihnea.Saracin@windriver.com>