Introduce SX to DX migration playbook
This commit introduces a migrate-subcloud.yml playbook that the user
can run, with an overrides file to provide config values, to perform
the migration steps for a subcloud. Once the migration playbook has
been applied and the subcloud has recovered from the unlock (performed
by the playbook), the second controller can be installed and
configured.

Story: 2008587
Task: 41743
Depends-On: https://review.opendev.org/c/starlingx/config/+/776536
Signed-off-by: Don Penney <don.penney@windriver.com>
Change-Id: I1d8c1219694147baaabb183ef3debe1715aaf153