Add a rehome subcloud playbook
This commit creates a rehome_subcloud playbook to migrate an existing
subcloud from the original central cloud to a new central cloud. This
playbook is expected to be remotely play from the new central cloud
system controller based on the subcloud's overrides. This playbook
updates the system controller network info in the subcloud, migrates
the keystone data and updates the subcloud's ca and certs. After
running this playbook and unlock the subcloud controllers, the
subcloud can be discovered online in the new central cloud, and
can be managed and brought in-sync from the new central cloud.

Test:
Successfully migrate a AIOSX and a AIODX subcloud to the new central
cloud with this playbook.

Depends-On: https://review.opendev.org/c/starlingx/config/+/784767
Depends-On: https://review.opendev.org/c/starlingx/stx-puppet/+/785977
Depends-On: https://review.opendev.org/c/starlingx/config/+/786638
Depends-On: https://review.opendev.org/c/starlingx/config/+/787213

Story: 2008774
Task: 42152
Change-Id: Iaa6699951e855c76602bd43f71d42e64e298c786
Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>