Fix kube-apiserver pod removal
This commit adds rolebinding configuration to bind
the "privileged-psp-user" role to the kubelet user.
It fixes the issue where the kube-apiserver pod does
not get recreated after enabling PodSecurityPolicy
plugin. With this fix we make sure to allow the apiserver
pod creation by granting permission to the kubelet user
to create that pod.

Closes-Bug: 1881605

Signed-off-by: Carmen Rata <carmen.rata@windriver.com>
Change-Id: Ibdf6d4cacf2ce83dfa744455dac460027b2a6e47