Remove creating admin endpoint cert in subcloud bootstrap
This change removes creating admin endpoint cert in subcloud
bootstrap.
The admin endpoint cert is generated in manifest at the time when
the controller node is unlock the first time. The cert data is
retrieved directly from k8s secret data (where cert-manager is
responsible to maintain it and keep it up to date).

Partial-Bug: 1923510
Depends-On: https://review.opendev.org/c/starlingx/stx-puppet/+/786666
Signed-off-by: Bin Qian <bin.qian@windriver.com>
Change-Id: Ie6a5c8fe159efcdebdb4c81666e981772408b82c