Fix restore user images playbook
When running restore_user_images playbook,
after the images will be uploaded to the local registry,
the docker images stored in the local docker filesystem
will be deleted by a python script called by the playbook.
Sometimes this script will try to delete images
that are not present into the local docker filesystem and
will throw some errors.
We fix this by adding a check before trying to delete an image.

Also, updated the
/playbooks/roles/common/push-docker-images/files/download_images.py
file to do the same checks since it does similar steps
and can also fail when deleting
the images from the local docker cache.

Closes-Bug: 1928092
Signed-off-by: Mihnea Saracin <Mihnea.Saracin@windriver.com>
Change-Id: I1be45b2669ad1ab209aa49f050106dd1a7759cee