Fix: Upgrade does not replay correctly after activate failure
When the etcd runtime manifest fails during upgrade activate,
the activation fails as expected. However when the upgrade-activate
is attempted again, the etcd upgrade playbook is not re-run.
Depending on when the manifest fails, this could result in secure
etcd remaining disabled on the system.

This commit makes it possible to replay the etcd upgrade playbook,
thus solving the issue.

Change-Id: I7f453d9040916381519ac96ed4567ec5fb6e7a8d
Signed-off-by: Jessica Castelino <jessica.castelino@windriver.com>
Closes-Bug: 1928130