Use async_timeout for kubectl wait timeout
Data migration failed for 5 subclouds during 30 subcloud upgrade.
This failure was caused by timeout waiting for Kubernetes component,
Networking or Armada pods to be ready.

The async_timeout is set to 120s in restore mode, but the pod wait time
is still set to 30s. This update changes the pod wait time to be the
same as async_timeout.

Tested on DC-2 with 50 subcloud parallel upgrade

Closes-Bug: 1928252

Signed-off-by: Tao Liu <tao.liu@windriver.com>
Change-Id: If2b21d0cd2e0de9e84869323a43d4a249d031132