Cleanup of subcloud rehoming playbook
As the ansible log is accessible for all the users, we don't want to
expose the keystone IDs and passwords in a subcloud's ansible log.
This commit hides the keystone IDs and passwords in the ansible log.

This commit also removes un-used facts and the rehome_in_progress
flag.

Tested with successfully re-homed a SX and a DX subcloud in virtual
box env.

Story: 2008774
Task: 42462

Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>
Change-Id: I6c2e83ba5b4923e9d7c82ccf94165608739e59e1