Fix deletion of temp dir when backup fails
When the /opt/backups disk is full and the backup playbook
is executed, the Ansible task fails to create a temp directory.
After the failure the cleanup code is executed but if the temp
dir was not created, it fails because tempdir.path will be undefined.
This second failure aborts the cleanup code and the
"System Backup in progress." will stay on indefinitely, requiring
manual intervetion.

Tested by executing the backup procedure when /opt/backups
is full.

Closes-Bug: 1928365
Signed-off-by: Isac Souza <IsacSacchi.Souza@windriver.com>
Change-Id: I76c852b65bde4b40a22bcb9be3a81776ade86d15