Configure kubeadm to not apply the default taint
The taint "node-role.kubernetes.io/master:NoSchedule" needs
to be removed from master node so that pods can be scheduled
on it. This is handled by a bootstrap task. However, issue
was seen that the default taint was not removed during bootstrap
that causes armada pod fails to be scheduled on controller-0.
This happens on one of the subcloud when bootstrapping a batch
of 50 subclouds.

Add configuration in kubeadm to not apply the default taint
at the beginning so it doesn't need to be removed afterwards.

Tested AIO-SX, DX upgrade and a batch of 50 subclouds deployment

Change-Id: I543280ddd55ec94ccf0586dc07877349baa06bdd
Closes-Bug: 1928722
Signed-off-by: Angie Wang <angie.wang@windriver.com>