Bug fix: update barbican external id if the project id changes
The previous commit
https://review.opendev.org/c/starlingx/ansible-playbooks/+/790824
unexpectedly moves the update_barbican_project_external_id out of the
scope of the project id changes. Further, the old_id['id'] is
duplicated with the current_project_id, it will be undeclared in the
case of the project id not changed (by deleting the subcloud in
powered off status and re-add it with --migrate option). This commit
deletes the duplicated variable and moves the
update_barbican_project_external_id to the correct scope.

Tests:
1 Delete a subcloud from a central cloud and add it back with the
migrate option.
2 Migrate a subcloud successfully to a different central cloud.
3 Migrate a subcloud successfully with an extra nonlocal user to a
different central cloud.

Partial-Bug: 1928139
Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>
Change-Id: I8acf7bcea5d8e77b92877427ccd40b95e8b3515e