Update armada image version to 7ef4b86
After the merge of [1] we are still seeing some instances of the armada
waiting indefinitely on Evicted pods until stx-openstack fails to apply.
Investigating this further, I figured that the armada pods were me
still using the old armada version due to the version of the images
declared on these ansible playbooks.

This commit upversion the Armada image to 7ef4b86 and Tiller to 2.16.9.

Tested in a custom build in conjunction with [2], I was able to
correctly deploy StarlingX and apply stx-openstack.

[1] https://review.opendev.org/c/starlingx/integ/+/783083
[2] https://review.opendev.org/c/starlingx/platform-armada-app/+/793972

Depends-On: I8532036cd70b3e20a143dceba19119779413eb39
Closes-bug: #1928994
Signed-off-by: Thiago Brito <thiago.brito@windriver.com>
Change-Id: Id06d805229b55a471a52c3173b2aa8b62957fcf9