Apply puppet manifest to re-configure/reload related services
After migrating the keystone IDs and passwords, related services are
expected to be re-configured and reloaded. This commit removes the
re-config and reload tasks in the playbook, instead applies a
subcloud_rehome puppet manifest to finish these tasks.

Test:
Migrate a SX subcloud(IPv4) and a DX subcloud(IPv6) to a new central
cloud with this playbook. The playbook completes successfully, after
that the subclouds can get managed and in-sync. The subclouds can
access the central registry after the migration.

Depends-On: https://review.opendev.org/c/starlingx/stx-puppet/+/795113
Story: 2008774
Task: 42582
Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>
Change-Id: Icf23d06478672d1f77dda3ff6475279aef089667