Check and add remote host key into known_hosts
The previous method of updating SSH known hosts removes the existing
host key of the ansible remote host. This commit changes this method
to scan and add the new ssh key of the ansible remote host into
~/.ssh/known_hosts. After this change, the ssh connection to the
ansible remote host will be accepted if the ansible remote host has
either the former key or the latter key.

Test see in:
https://review.opendev.org/c/starlingx/ansible-playbooks/+/798918

Partial-Bug: 1934154
Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>
Change-Id: I7c56e130ad8b75f5c8ffead0b2041a5dedac6343