Fix backup remote play
Script module fails to find a binary on the Ansible controller.

The correct module to be used is 'command' in the first place, since the
binary is installed at a predifined location and not bundled with the
playbooks themselves.

Tested local play.
Tested remote play from a non-StarlingX Ansible controller.

Closes-Bug: 1934754
Signed-off-by: Dan Voiculeasa <dan.voiculeasa@windriver.com>
Change-Id: I8a6209cf47bc84437bd9535aa5e70611cc093579