Platform.conf has duplicated region_config entries
Bootstrap playbook is currently writing two times the key, first by
puppet and second on the task "Add new entries to platform.conf". This
removes the second addition.

Tested running the playbook on a SX, DC, on subcloud Backup & Restore,
and re-ran it on the subcloud. In all cases the platform.conf file
contains only one region_config and the subclouds were in sync with no
unexpected alarms.

Closes-bug: #1931529
Change-Id: I55f17ddfdedf5cc6b16c7d55daa99e57bfbb38bc
Signed-off-by: Iago Regiani <Lago.RodriguezRegiani@windriver.com>