Add support for selecting the k8s version on install
On initial install we want to be able to specify the desired kubernetes
version via the localhost.yml file. The suggested format is something
like this:

kubernetes_version: 1.18.1

At initial install time if the version is specified in the localhost.yml
file the desired version will be written to the sysinv DB (overriding
the default) by the bootstrap Ansible playbook.

The kubernetes_version specified in the localhost.yaml will be
propagated to the kubeadm.yaml.erb file.

Also at initial install time since we call kubeadm and
kubectl commands in the ansible playbooks, we need to ensure that the
/usr/local/kubernetes/current/stage1 and
/usr/local/kubernetes/current/stage2 bind mounts
are set to the desired version for the initial install.
If the desired version is 1.18.1, we will bind-mount
/usr/local/kubernetes/1.18.1/stage1
onto /usr/local/kubernetes/current/stage1 and
/usr/local/kubernetes/1.18.1/stage2
onto /usr/local/kubernetes/current/stage2.

Story: 2008972
Task: 42913
Depends-On: Ic848ea777fecab5bda7488df225c2c3e924c2fd5
Change-Id: Ie1bc09dfd7ccdf87633acf1db6d1112af6b78f31
Signed-off-by: Mihnea Saracin <Mihnea.Saracin@windriver.com>