Support remote playback of upgrade playbooks
The following commit introduced a change to run the
upgrade-k8s-networking role during the upgrade_platform playbook. When
running the upgrade_platform remotely, incorrect data was included by
the upgrade-k8s-networking role. The distributed_cloud_role was set to
systemcontroller, as the values were read in from the system controller,
not the sublcoud.

https://opendev.org/starlingx/ansible-playbooks/commit/9addaf9a34

The solution is to copy the overrides_file to the local filesystem when
running the role/playbook on a remote target. All upgrade playbooks have
been updated to avoid this issue in the future.

Tests:
Remote and local playback of the following playbooks
* upgrade_platform
* upgrade-k8s-networking
* upgrade-static-images
* upgrade-k8s-armada-helm

Closes-Bug: 1939161
Change-Id: Id3fdf802e63d2f1c9947f265b07ea1159fa029ce
Signed-off-by: David Sullivan <david.sullivan@windriver.com>