Fix k8s 1.19 match selectors for SR-IOV images
When performing a k8s upgrade from 1.18.1 to 1.19.13, the
networking images will be upgraded as well.  As part of this
upgrade, the respective kubernetes spec templates which
align with the new images will be used.

An issue has been seen in which the rolling upgrade of the
SR-IOV daemonsets fail because the latest templates specify
a 'name' as the matchLabel selector.  However, the older spec
uses 'app' and 'tier' as the match selector.

I believe that as of apps/v1, the selector label is still
immutable for controllers such as daemonsets that have
already been deployed.

Some discussion on the topic can be found here:

https://github.com/kubernetes/kubernetes/issues/50808

For now, we'll just carry forward the 1.18 match selector.
It's possible this can be fixed in later k8s releases.

Story: 2008972
Closes-Bug: 1942351

Signed-off-by: Steven Webster <steven.webster@windriver.com>
Change-Id: Id0ca32038dc2897879786a17f9794515457cd837