Restore admin password during bootstrap manifest
For restore operations (including SX upgrades) we need to use the
updated admin password. The admin password can be changed after the
system is bootstrapped. In which case the password is stored in
secure_system instead of secure_static hieradata. When we use the old
admin password the bootstrap playbook will fail as the keyring password
does not match what is in the keystone database.

To correct this, we copy the password to the runtime hieradata used to
run the bootstrap manifest.

Tests run:
AIO-SX Bootstrap + Deploy
AIO-SX Backup and Restore
AIO-SX Upgrade

Closes-Bug: 1942907
Change-Id: I7549b846a470dff272ab17fc610ebe06260f439b
Signed-off-by: David Sullivan <david.sullivan@windriver.com>