Automatically select k8s version during bootstrap
When doing a fresh install (i.e. when not doing an upgrade) we
will eventually want to use the newest version of Kubernetes
available in the load.

For now however we just want to get some extra testing on K8s 1.19
so we're going to make it the default for fresh installs.

For the actual release we'll be defaulting to 1.21.

When doing an upgrade of AIO-SX however the DB is not restored
until later, so we need to also specify the Kubernetes version
to use for upgrades.  This will need to be updated every release
unless we can figure out a better alternative.

Story: 2008972
Task: 43468
Signed-off-by: Chris Friesen <chris.friesen@windriver.com>
Change-Id: I59873ef9efd949e80218affa3bcca24ae6175e01