Set toleration overrides on armada-overrides.yaml
A toleration needs to be added to all resources that create pods since
the node-role.kubernetes.io/master taint will be restored to all master
nodes. This ensures that the pods will be scheduled and run on
the master node.

This change adds the toleration for the
node-role.kubernetes.io/master:NoSchedule taint on
armada-overrides.yaml.j2.

Test plan:

- Building and installing a new iso with this patch and the taint
applied. The armada-api pod was able to run on the tainted node.
- Confirmed that toleration is present when describing the armada pod.

Depends-On: https://review.opendev.org/c/starlingx/integ/+/811793
Story: 2009232
Task: 43546

Signed-off-by: Enzo Candotti <enzo.candotti@windriver.com>
Change-Id: I7308f27d890c500afb8ea828907483bc2a59a4bb