Update registry credentials during rehoming
As we are switching to use 'sysinv' user instead of 'admin' user to
access the registries, this commit adds a task in the rehoming
playbook to update the registry credentials with the sysinv
credentials from the new system controllers which a subcloud is
migrating to.

Test steps:
1. Deploy a AIOSX subcloud in central cloud A, update the subcloud's
registries with its sysinv credentials.
2. Update the admin credentials from central cloud B in the subcloud.
3. Migrate the subcloud to central cloud B.
4. Lock/unlock the subcloud after its deploy status turns to
"complete" state.

Test result:
The subcloud turns online after unlocking and turns to "in-sync" after
being managed by central cloud B. The registries auth-secrets are all
updated to sysinv credentials from central cloud B. The central
registry can be accessed from the subcloud with the sysinv user and
its password.

Depends-On: https://review.opendev.org/c/starlingx/utilities/+/814645
Closes-Bug: 1947014
Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>
Change-Id: I384930d3842f8a4da03648af7153dea430c49baa