Limit tasks for upgrade on common/armada-helm
This change adds restrictions to each task that it's not needed when
upgrading Armada Helm Chart.

This is done by adding the "when" statement, evaluating the mode type.
On upgrades, the mode is "upgrade_k8s_armada_helm".

Test Plan:

PASS: Build a designer iso with an upgrade script for armada charts and
      run an upgrade.
      Confirmed that the executed tasks were the expected.
PASS:  Build a designer iso and run a fresh install.

Story: 2009232
Task: 43647

Signed-off-by: Enzo Candotti <enzo.candotti@windriver.com>
Change-Id: I3ee13f394e2961f95040f4360549f800b6e2d533