Fix restore playbook for restore-more-data
In the restore-more-data playbook, the ceph crushtool tool is called
to remove the osds from crushmap backup file.
For the mimic version, it is possible to call the same file for input
and output with no errors.
For the nautilus version, the input and output files should not be the
same to avoid errors.

To fix this a temporary file is set as output of the crushtool command
and then rename this file back to the original name.

Test Plan:

PASS: Run a backup/restore successfully.

Story: 2009074
Task: 43869
Signed-off-by: Felipe Sanches Zanoni
<Felipe.SanchesZanoni@windriver.com>
Change-Id: I09b39dc4a85d73b7da39198e6a925008353d6ebe