Reconfigure ETCD when root CA filenames change
This change adds support to reconfigure, in a bootstrap replay, the ETCD
when the root CA filenames change.

Test Plan:

PASS: Create an AIO-SX using a new ISO created with this change, run
bootstrap successfully, then run it again with a different root CA files
and check that the ETCD certs are correctly generated.

Closes-Bug: 1949209
Signed-off-by: Joao Victor Portal <Joao.VictorPortal@windriver.com>
Change-Id: Icc7152185e4c4d0440c8f05ffe7cf9dc78f3795b