Apply apiserver parameters on restore and upgrade
The issue is manifested by backing-up and restoring a system with
kube-apiserver parameters applied, e.g.: oidc-auth params. The
parameters are restored and can be displayed by running
`service-parameter-list` but aren't getting reapplied back.

The fix consists in running the `puppet-manifest-apply.sh` script during
the 'restore-more-data' step of the restore, which is executed both
during platform restore and upgrade. When applying the manifest the
kube-apiserver service parameters are updated, similarly to running
`service-parameter-apply`.

Test Plan: In an AIO-SX, apply oidc-auth kube-apiserver parameters,
perform a B&R and an upgrade operation and verify these are properly
restored.

PASS: Verify kube-apiserver service parameters are restored after a
System Backup and Restore
PASS: Verify kube-apiserver service parameters are restored after a
System Upgrade

Closes-Bug: 1948140
Signed-off-by: Rafael Camargos <RafaelLucas.Camargos@windriver.com>
Change-Id: I9d4823deb0b928acd86c8372fe5306ac8a841518