Update B&R monitor bootstrap to enable V1/V2 Msgr protocols
During system restore, the monitor does not start up currently and hangs
in a continual probing loop.

The issue here stems from the fact that we are no longer providing the
port for the public_addr of the Ceph monitor. This was removed with
https://opendev.org/starlingx/config/commit/09e0401d59d94cf619f3810244c5ff882f6a9213
and is not present the restored ceph.conf.

If we define a port, either v1 (6789) or v2 (3300), for the monitor's
public_addr, then the restore playbook will work without us needing to
change the backup.

Since we still have older containerized ceph clients using the V1
protocol, we can't specify the V2 port because we'll have no easy
way to enable the V1 protocol.

Here we add the V1 protocol port to allow the monitor to bootstrap
correctly then enable the V2 protocol.

Test plan:
 - AIO-SX B&R
 - AIO-DX B&R

Story: 2009074
Task: 43992
Signed-off-by: Felipe Sanches Zanoni <Felipe.SanchesZanoni@windriver.com>
Change-Id: I487035f96654371f708d846803238de1ad995572