Force calico to use NFT iptables backend
This commit sets a calico environment variable to force the use of
iptables-nft in the calico-node pod.

StarlingX recently uprevved iptables on the host to 1.8.4.  An
implication of this change is that iptables is now using the nftables
kernel API and infrastructure to store firewall / traffic rules.
For example, running iptables --version on the latest release will
show:

iptables v1.8.4 (nf_tables)

This indicates that by default, nftables is being used, rather than
the older iptables-legacy infrastructure.
In fact, StarlingX does not even provide the user-space mechanism to
see the iptables-legacy tables anymore.
It can however be seen via a pod with the iptable-legacy binary,
such as via the calico-node pod.

nf_tables is the right way to move forward for us, as it has a
number of benefits over the older iptables (legacy) infrastructure.

One thing to note is that it's entirely possible that a
pod/container can have it's own version of the user-space iptables.
That is, a pod based on an older iptables may only be programming
the legacy tables, while a newer one may be using the iptables-nft
version. This seems to be causing the mismatch, although
theoretically the underlying xtables datapath _should_ be the same
between the two.

It has been seen that the move to iptables 1.8.4 has caused an
occasional loss of connectivity between pods.
Specifically, the traffic seems to be caught up in the FORWARD chain
of the iptables filter table.  Because of a missing cali-FORWARD rule,
the traffic is dropped as per the default policy of the chain. the
cali-FORWARD rule was found to be in the legacy table, rather than the
nft table.

Currently there are three k8s components that modify iptables -
kubelet, kube-proxy, and calico-node. They all seem to share the same
algorigthm for determining whether to use iptables-nft, or
iptables-legacy. Basically, the number of rules from each table is
taken, and whichever is higher wins.  This seems to be working for k8s
(host), and kube-proxy.  However, we've seen cases where Calico is
using the iptables-legacy backend.

This commit tells calico explicitly to use the NFT backend.

Note 2: The issue was actually only seen with k8s 1.18, which has
Calico 3.12. Later versions of Calico did not exhibit the same
behaviour, but I'm changing the k8s 1.19+ versions to explicitly set
the iptables backend in Calico as well.
Since StarlingX does not even have a mechanism for manipulating
the iptables-legacy tables from user-space anymore, I think it's
safer moving forward to set the backend to "NFT".

Test Plan:

Verifying iptables-nft and iptables-legacy via calico-node pod with
fresh install:

AIO-SX w/ k8s 1.18
AIO-SX w/ k8s 1.19
AIO-SX w/ k8s 1.21
Standard w/ k8s 1.18
Standard w/ k8s 1.21

Verifying host <-> pod, pod <-> pod, service communication:

AIO-SX w/ k8s 1.21

Verifying host-upgrade from release with older iptables (legacy) to
StarlingX master (iptables 1.8.4 with nft)

AIO-SX

Closes-Bug: 1952351
Change-Id: I0dd7e14cfdce21c5ce9127d17fc6914c4c118435
Signed-off-by: Steven Webster <steven.webster@windriver.com>