Revert "Apply apiserver parameters on restore and upgrade"
This reverts commit 3d2eb4e267461614f9c0ae897e058689812d9208.

Reason for revert: During tests, it was detected that the 'Apply puppet
apiserver parameters' restore step is broken on standard systems.
Initial investigations showed that the issue only happens on standard
systems and can be related to the hieradata structure, so this should be
reverted until we have a proper fix for it.

Change-Id: I5bba441b99318be6341469a41ad98e49009d3172