Use formatted IP for monitor's public_addr in ceph.conf
It is needed to set ceph v1 protocol port for the public_addr key inside
section mon.controller (AIO-DX) or mon.controller-0.
It is correctly managing ipv4 addresses but is failing to manage ipv6
addresses.

The fix consists in getting the ip address from mgr, detecting if it is
ipv4 or ipv6. If it is ipv6 wraps the ip in sqaure brackets. Then it
adds the port number for ceph v1 protocol.

This way it is possible to bring up ceph monitor during restore process
for ipv4 and ipv6 setups.

Test Plan:
 - B&R AIO-SX ipv6
 - Upgrade 21.05 -> 21.12 AIO-SX ipv6

Task: 44044
Story: 2009074
Signed-off-by: Felipe Sanches Zanoni <Felipe.SanchesZanoni@windriver.com>
Change-Id: Ic1374d28be86bb6c3434fa913b8fe70575b6dbe4