Fix for Calico's use of iptables on multi-node systems
Commit 43af74a27f3173700d7958fc52a23f3ec96150c2 introduced the
explicit setting of the FELIX_IPTABLES="NFT" environment
variable to the calico-node pods to get around an issue seen
with calico using the legacy iptables.  This worked for fresh
installs and backup-restore or upgrades of an AIO-SX system.

It has been discovered that this causes problems for upgrades
in AIO-DX/Simplex systems.  When the first controller is upgraded
in a system, it will still be running the old Calico 3.12 until
such time as it is swacted to.  Since the previous release did
not explicitly set the FELIX_IPTABLESBACKEND, the default value
for this will be used, even if an iptables backend of NFT is
detected.

In Calico 3.19, which is present in current/upgraded loads,
the default value of the iptables backend was changed to 'auto',
which means the auto-detection method for determining which
iptables to use is triggered unless explicitly set to one of
'legacy' or 'NFT'.

Once the system has been swacted to, the value of the backend
is explicitly set to "NFT", and the daemonset is re-deployed.
This causes the upgraded host to then have both legacy and
NFT rules present, and the un-upgraded host to also be
instructed to use NFT, even though it's underlying iptables
is the legacy version.

To address this limitation, this commit explicitly sets the
FELIX_IPTABLESBACKEND to 'auto' such that the detection
mechanism can be triggered on hosts still running iptables
legacy or the newer nf_tables implementation.

In addition, during upgrade testing on an IPv6 system,
it has been found that there is a potential for Calico's
default IPv4 pool to overlap with the subnet specified
for the IPv4 pxe network.  Although we don't explicitly set
the CALICO_IPV4POOL_CIDR in an IPv6 system, Calico will
still create this by default, using the default:
[First not used in locally of (192.168.0.0/16, 172.16.0.0/16,
.., 172.31.0.0/16) ]

Note: In an IPv4 system, a default IPv6 pool is not created
in the same way.

This can cause the cali-nat-outgoing iptables rule to be
triggered for pxeboot traffic, causing the source port of
DHCP server responses to be randomized.  Some PXE clients
will reject this traffic.

To address this limitation, this commit explicitly creates
appropriate ippools, rather than relying on the setting of
the calico-node environment variables.  On an upgrade, the
default-ipv4-ippool will be deleted if it is detected that
the system is IPv6.

Test plan:
PASS  AIO-SX bootstrap install for an IPv4 system
PASS  AIO-SX bootstrap install for an IPv6 system
PASS  AIO-DX bootstrap install for an IPv4 system
PASS  AIO-DX bootstrap install for an IPv6 system
PASS  AIO-DX upgrade for an IPv4 system
PASS  AIO-DX upgrade for an IPv6 system

Closes-Bug: 1953232

Signed-off-by: Andre Fernando Zanella Kantek <AndreFernandoZanella.Kantek@windriver.com>
Change-Id: I1a611988cd45e03060d08256950552bed0fe64f5