Fix root partition detection for ostree
The /proc/cmdline behaves a bit differently for ostree based
systems. The hard disk is partitioned for an A/B setup so if
the user takes a snapshot of their system and something goes
wrong then they can rollback their system to an earlier state.
As a result the /proc/cmdline is setup differently from a vanilla
machine.

In order to identify the root disk of an ostree based system we
check for ostree_root's disk label, once that is determined we
check the symlink for /dev/disk/by-label for the disk device.

This fixes a failure when the playbook runs because the
check_root_disk_size.py doesnt understand about ostree.

Test Plan:
PASS Apply patch
PASS Boot Centos ISO check to ansible playbook log to make sure the
script was able to detect root partition.
PASS Boot Debian ISO check to see an exception was not raise in
the ansible playbook log.

Story: 2009101
Task: 44279

Signed-off-by: Charles Short <charles.short@windriver.com>
Change-Id: Iaadbea94e91eb438084705fdd45c165c9847f87e